requirements.txt
click==8.0.1
progressbar2==3.53.2
psutil==5.8.0
python-utils==2.5.6
six==1.16.0
tqdm==4.62.2
tomli==2.0.1